Corrected the credentials selection flow
The credential selection now works like this:
Check if there is a credential on the current app submission, if not, check a previous submission for this app. If we still don't have a credential, get a credential from a past submission on any app.
If there isn't a credential you must generate or upload a certificate.